Refactoring Tabs Requests and Tabs
