+++ resources/js/processes-catalogue/components/ProcessTab.vue
@@ -5,76 +5,32 @@
         title="My Requests"
         active
       >
-         <div class="bg-white" v-if="showTabRequests">
-          <filter-table
-            :headers="tableHeadersRequests"
-            :data="dataRequests"
-            @table-row-click="(row) => openRequest(row)"
-          />
-          <pagination-table
-            :meta="dataRequests.meta"
-            @page-change="changePage"
-          />
-        </div>
-        <div v-else>
-          <default-tab
-            :alt-text="$t('No Image')"
-            :title-text="$t('You have made no requests of this process.')"
-            :description-text="$t('All your requests will be shown here')"
-          />
-        </div>
+        <request-tab
+          :currentUser="currentUser"
+          :process="process"
+        ></request-tab>
       </b-tab>
-      <b-tab class="bg-white" title="My Tasks">
-        <div v-if="showTabTasks">
-          <filter-table
-            :headers="tableHeadersTasks"
-            :data="dataTasks"
-            @table-row-click="(row) => openTask(row)"
-          />
-          <pagination-table
-            :meta="dataTasks.meta"
-            @page-change="changePage"
-          />
-        </div>
-        <div v-else>
-          <default-tab
-            :alt-text="$t('No Image')"
-            :title-text="$t('You have no tasks from this process')"
-            :description-text="
-              $t('All your tasks related to this process will be shown here.')
-            "
-          />
-        </div>
+      <b-tab
+        class="bg-white"
+        title="My Tasks"
+      >
+        <task-tab
+          :currentUser="currentUser"
+          :process="process"
+        ></task-tab>
       </b-tab>
     </b-tabs>
   </div>
 </template>
 
 <script>
-import Vue from "vue";
-import AvatarImage from "../../components/AvatarImage";
-import PMColumnFilterPopover from "../../components/PMColumnFilterPopover/PMColumnFilterPopover.vue";
-import paginationTable from "../../components/shared/PaginationTable.vue";
-import TasksList from "../../tasks/components/TasksList.vue";
-import DefaultTab from "./DefaultTab.vue";
-import isPMQL from "../../modules/isPMQL";
-import ListMixin from "../../requests/components/ListMixin";
-import { FilterTable } from "../../components/shared";
-import moment from "moment";
-import { createUniqIdsMixin } from "vue-uniq-ids";
-
-const uniqIdsMixin = createUniqIdsMixin();
-
-Vue.component("AvatarImage", AvatarImage);
-
+import RequestTab from "./RequestTab.vue";
+import TaskTab from "./TaskTab.vue";
 export default {
   components: {
-    PMColumnFilterPopover,
-    paginationTable,
-    TasksList,
-    DefaultTab,
+    RequestTab,
+    TaskTab,
   },
-  mixins: [uniqIdsMixin, ListMixin],
   props: {
     currentUser: {
       type: Object,
@@ -83,288 +39,9 @@ export default {
       type: Object,
     },
   },
-  data() {
-    return {
-      pmqlRequest: `(status = "In Progress") AND (requester = "${this.currentUser.username}")`,
-      pmqlTask: `(status = "In Progress")`,
-      filter: "",
-      previousFilter: "",
-      previousPmql: "",
-      orderBy: "id",
-      orderDirection: "DESC",
-      additionalParams: "",
-      advanced_filter: "",
-      sortOrder: [
-        {
-          field: "id",
-          sortField: "id",
-          direction: "desc",
-        },
-      ],
-      fields: [],
-      showTabRequests: false,
-      showTabTasks: false,
-      tableHeaders: [],
-      tableHeadersRequests: [
-        {
-          label: "CASE #",
-          field: "case_number",
-          sortable: true,
-          default: true,
-          width: 55,
-        },
-        {
-          label: "CASE TITLE",
-          field: "case_title",
-          sortable: true,
-          default: true,
-          width: 220,
-        },
-        {
-          label: "STATUS",
-          field: "status",
-          sortable: true,
-          default: true,
-          width: 150,
-        },
-        {
-          label: "STARTED",
-          field: "initiated_at",
-          format: "datetime",
-          sortable: true,
-          default: true,
-          width: 160,
-        },
-        {
-          label: "COMPLETED",
-          field: "completed_at",
-          format: "datetime",
-          sortable: true,
-          default: true,
-          width: 160,
-        },
-      ],
-      tableHeadersTasks: [
-        {
-          label: "CASE TITLE",
-          field: "case_title",
-          sortable: true,
-          default: true,
-          width: 140,
-          truncate: true,
-        },
-        {
-          label: "PROCESS NAME",
-          field: "name",
-          sortable: true,
-          default: true,
-          width: 140,
-          truncate: true,
-        },
-        {
-          label: "TASK NAME",
-          field: "element_name",
-          sortable: true,
-          default: true,
-          width: 140,
-          truncate: true,
-        },
-        {
-          label: "STATUS",
-          field: "status",
-          sortable: true,
-          default: true,
-          width: 100,
-        },
-        {
-          label: "DUE DATE",
-          field: "due_at",
-          format: "datetime",
-          sortable: true,
-          default: true,
-          width: 140,
-        },
-      ],
-      dataRequests: {},
-      dataTasks: {},
-      savedSearch: false,
-      queryTask: "",
-      queryRequest: "",
-      perPage: 10,
-    };
-  },
-  mounted() {
-    this.queryBuilder();
-  },
-  methods: {
-    jsonRows(rows) {
-      if (rows.length === 0 || !_.has(_.head(rows), "_json")) {
-        return rows;
-      }
-      return rows.map((row) => JSON.parse(row._json));
-    },
-    changePage(page) {
-      this.page = page;
-      this.queryBuilder();
-    },
-    openRequest(data, index) {
-      return `/requests/${data.id}`;
-    },
-    transform(data) {
-      // Clean up fields for meta pagination so vue table pagination can understand
-      data.meta.last_page = data.meta.total_pages;
-      data.meta.from = (data.meta.current_page - 1) * data.meta.per_page;
-      data.meta.to = data.meta.from + data.meta.count;
-      data.data = this.jsonRows(data.data);
-      console.log("EN TRANSFORM: ", data.data);
-      for (let record of data.data) {
-        //format Status
-        record["case_number"] = this.formatCaseNumber(record);
-        record["status"] = this.formatStatus(record["status"]);
-        record["participants"] = this.formatParticipants(record["participants"]);
-      }
-      return data;
-    },
-    formatParticipants(participants) {
-      return {
-        component: "AvatarImage",
-        props: {
-          size: "25",
-          "input-data": participants,
-          "hide-name": false,
-        },
-      };
-    },
-    openTask(task) {
-      return `/tasks/${task.id}/edit`;
-    },
-    formatStatus(status) {
-      let color = "success",
-      label = "In Progress";
-      switch (status) {
-        case "DRAFT":
-          color = "danger";
-          label = "Draft";
-          break;
-        case "CANCELED":
-          color = "danger";
-          label = "Canceled";
-          break;
-        case "COMPLETED":
-          color = "primary";
-          label = "Completed";
-          break;
-        case "ERROR":
-          color = "danger";
-          label = "Error";
-          break;
-      }
-      return (
-        '<span class="badge badge-' +
-        color +
-        ' status-' +
-        color +
-        '">' +
-        this.$t(label) +
-        "</span>"
-      );
-    },
-    formatCaseNumber(value) {
-      return `
-      <a href="${this.openRequest(value, 1)}"
-         class="text-nowrap">
-        # ${value.case_number}
-      </a>`;
-    },
-    queryBuilder() {
-      let pmql = "";
-      if (this.pmqlRequest !== undefined) {
-        pmql = this.pmqlRequest;
-      }
-      let filter = this.filter;
-      if (filter && filter.length) {
-        if (filter.isPMQL()) {
-          pmql = `(${pmql}) and (${filter})`;
-          filter = "";
-        }
-      }
-      if (this.previousFilter !== filter) {
-        this.page = 1;
-      }
-      this.previousFilter = filter;
-      if (this.previousPmql !== pmql) {
-        this.page = 1;
-      }
-      this.previousPmql = pmql;
-      this.tabRequests();
-      this.tabTasks();
-    },
-    tabRequests() {
-      this.queryRequest =
-        "requests?page=" +
-        this.page +
-        "&per_page=" +
-        this.perPage +
-        "&include=process,participants,activeTasks,data" +
-        "&pmql=" +
-        `${this.pmqlRequest}` +
-        " AND process_id=" +
-        `${this.process.id}` +
-        "&filter&order_by=id&order_direction=DESC";
-      this.getData(this.queryRequest, "requests");
-    },
-    tabTasks() {
-      this.queryTask =
-        "tasks?page=" +
-        this.page +
-        "&include=process,processRequest,processRequest.user,user,data" +
-        "&pmql=" +
-        `${this.pmqlTask}` +
-        " AND process_id=" +
-        `${this.process.id}` +
-        "&per_page=10&order_by=ID&order_direction=DESC&non_system=true";
-      
-      this.getData(this.queryTask, "tasks");
-    },
-    getData(query, type) {
-    // Load from api client
-      ProcessMaker.apiClient
-        .get(query)
-        .then((response) => {
-          console.log('GET DATA: ', response.data);
-          const dataResponse = this.transform(response.data);
-          type === "requests"
-            ? (this.dataRequests = this.transform(response.data))
-            : type === "tasks" &&
-              (this.dataTasks = this.transform(response.data));
-          if (
-            dataResponse &&
-            Array.isArray(dataResponse.data) &&
-            dataResponse.data.length === 0
-          ) {
-            type === "requests"
-              ? (this.showTabRequests = false)
-              : (this.showTabTasks = false);
-          } else {
-            type === "tasks"
-              ? (this.showTabRequests = true)
-              : (this.showTabTasks = true);
-          }
-        })
-        .catch((error) => {
-          if (error.code === "ERR_CANCELED") {
-            return;
-          }
-          if (_.has(error, "response.data.message")) {
-            ProcessMaker.alert(error.response.data.message, "danger");
-          } else if (_.has(error, "response.data.error")) {
-            return;
-          } else {
-            throw error;
-          }
-        });
-    },
-  },
+  // data() {
+  //   return {};
+  // },
 };
 </script>
 <style>
